Added Payments fixes and seperated files
